+++ common/include/common/rfproto.h
@@ -165,7 +165,10 @@ enum RF_GameType
     RF_GT_DM     = 0x00,
     RF_GT_CTF    = 0x01,
     RF_GT_TEAMDM = 0x02,
-    RF_GT_KOTH = 0x03, // Supported in Alpine v1.2+, game_info packet delivered to legacy clients uses 0x02 instead to prevent them crashing
+    // Below types are supported in Alpine v1.2+, but game_info packets
+    // delivered to legacy clients uses 0x02 instead to prevent them crashing
+    RF_GT_KOTH = 0x03,
+    RF_GT_DC = 0x04,
 };
 
 enum RF_ServerFlags