+++ game_patch/multi/network.cpp
@@ -338,9 +338,9 @@ FunHook<MultiIoPacketHandler> process_game_info_packet_hook{
 CodeInjection process_game_info_packet_game_type_bounds_patch{
     0x0047B30B,
     [](auto& regs) {
-        // Valid game types are between 0 and 3
-        // 3 = KOTH
-        regs.ecx = std::clamp<int>(regs.ecx, 0, 3);
+        // Valid game types are between 0 and 4
+        // 3 = KOTH, 4 = DC
+        regs.ecx = std::clamp<int>(regs.ecx, 0, 4);
     },
 };